+++ data-raw/test_glmerb_airbnb.R
@@ -61,7 +61,7 @@ fit <- glmerb(
   form,
   data = dat,
   family = poisson(),
-  measurement_prior_list = ps,
+  pfamily_list = pfamily_list(ps),
   n = n_draw,
   seed = 42L
 )